requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.1
    # via django
cffi==1.15.0
    # via cryptography
cryptography==37.0.2
    # via -r requirements.in
damm32==1.2.1
    # via -r requirements.in
django==3.2.13
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-rest-registration
    #   django-stubs-ext
    #   django-treebeard
    #   djangorestframework
    #   djangorestframework-simplejwt
django-autoslug==1.9.8
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-rest-registration==0.7.2
    # via -r requirements.in
django-stubs-ext==0.5.0
    # via -r requirements.in
django-treebeard==4.5.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   django-rest-registration
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via djangorestframework-simplejwt
pytz==2022.1
    # via
    #   django
    #   djangorestframework
sqlparse==0.4.2
    # via django
typing-extensions==4.3.0
    # via django-stubs-ext
